requirements.txt
PyYAML==6.0
requests==2.27.1
python-dotenv==0.20.0